@@ -53,7 +53,7 @@ dependencies:
     path: packages/rust_sdk
   mime: ^1.0.2
   flutter_colorpicker: ^1.0.3
-  emoji_picker_flutter: ^1.3.0
+  emoji_picker_flutter: ^1.5.2
   filesystem_picker: ^2.0.1
   permission_handler: ^10.2.0
   transparent_image: ^2.0.0
