@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:effektio/controllers/chat_room_controller.dart';
+import 'package:effektio/models/JoinedRoom.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart'
     show
@@ -9,31 +10,23 @@ import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart'
         FfiListConversation,
         FfiListInvitation,
         Invitation,
-        RoomMessage,
         TypingEvent;
+import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:get/get.dart';
 
-//Helper class.
-class JoinedRoom {
-  Conversation conversation;
-  RoomMessage? latestMessage;
-  List<types.User> typingUsers = [];
-
-  JoinedRoom({
-    required this.conversation,
-    this.latestMessage,
-  });
-}
-
 class ChatListController extends GetxController {
   Client client;
+
   List<JoinedRoom> joinedRooms = [];
   List<Invitation> invitations = [];
+  bool showSearch = false;
+  List<JoinedRoom> searchData = [];
   bool initialLoaded = false;
   StreamSubscription<FfiListConversation>? _convosSubscription;
   StreamSubscription<FfiListInvitation>? _invitesSubscription;
   StreamSubscription<TypingEvent>? _typingSubscription;
+  TextEditingController searchController = TextEditingController();
 
   ChatListController({required this.client}) : super();
 
@@ -42,23 +35,52 @@ class ChatListController extends GetxController {
     super.onInit();
 
     _convosSubscription = client.conversationsRx().listen((event) {
-      if (!initialLoaded) {
-        initialLoaded = true; // used for rendering
-      }
       joinedRooms.clear();
       for (Conversation convo in event.toList()) {
         String roomId = convo.getRoomId();
         int pos = joinedRooms.indexWhere((x) {
           return x.conversation.getRoomId() == roomId;
         });
+
         JoinedRoom newItem = JoinedRoom(conversation: convo);
         if (pos == -1) {
           newItem.latestMessage = convo.latestMessage();
         } else {
+          if (joinedRooms[pos].avatar != null) {
+            newItem.avatar = joinedRooms[pos].avatar;
+          }
+          if (joinedRooms[pos].displayName != null) {
+            newItem.displayName = joinedRooms[pos].displayName;
+          }
           newItem.latestMessage = joinedRooms[pos].latestMessage;
           newItem.typingUsers = joinedRooms[pos].typingUsers;
         }
+
+        if (newItem.latestMessage != null) {
+          debugPrint(
+            'timestamp is ${newItem.latestMessage!.eventItem()!.originServerTs()}',
+          );
+        }
+
         joinedRooms.add(newItem);
+      }
+
+      joinedRooms.sort((a, b) {
+        if (a.latestMessage != null && b.latestMessage != null) {
+          return b.latestMessage!
+              .eventItem()!
+              .originServerTs()
+              .compareTo(a.latestMessage!.eventItem()!.originServerTs());
+        } else {
+          return 0;
+        }
+      });
+
+      joinedRooms.reversed;
+      searchData.addAll(joinedRooms);
+
+      if (!initialLoaded) {
+        initialLoaded = true; // used for rendering
       }
       update(['chatlist']);
     });
@@ -117,6 +139,38 @@ class ChatListController extends GetxController {
   void moveItem(int from, int to) {
     JoinedRoom item = joinedRooms.removeAt(from);
     joinedRooms.insert(to, item);
+    update(['chatlist']);
+  }
+
+  Future<void> setRoomProfile(Conversation room, JoinedRoom item) async {
+    await room.getProfile().then((value) {
+      if (value.hasAvatar()) {
+        item.avatar = value.getThumbnail(62, 60);
+      }
+      item.displayName = value.getDisplayName();
+    });
+  }
+
+  void searchedData(String data, List<JoinedRoom> listOfRooms) {
+    searchData.clear();
+    var name = '';
+
+    if (data.isNotEmpty) {
+      for (var element in listOfRooms) {
+        name = element.displayName!;
+        if (name.toLowerCase().contains(data.toLowerCase())) {
+          searchData.add(element);
+        }
+      }
+      update(['chatlist']);
+    } else {
+      searchData.addAll(joinedRooms);
+      update(['chatlist']);
+    }
+  }
+
+  void toggleSearchView() {
+    showSearch = !showSearch;
     update(['chatlist']);
   }
 }
