@@ -10,8 +10,7 @@ import 'package:effektio/screens/HomeScreens/chat/Overview.dart';
 // import 'package:effektio/screens/HomeScreens/Notification.dart';
 import 'package:effektio/screens/HomeScreens/faq/Overview.dart';
 import 'package:effektio/screens/HomeScreens/news/News.dart';
-import 'package:effektio/screens/HomeScreens/todo/AddToDo.dart';
-import 'package:effektio/screens/HomeScreens/todo/ToDoScreen.dart';
+import 'package:effektio/screens/HomeScreens/todo/ToDo.dart';
 import 'package:effektio/screens/OnboardingScreens/LogIn.dart';
 import 'package:effektio/screens/OnboardingScreens/Signup.dart';
 import 'package:effektio/screens/SideMenuScreens/Gallery.dart';
@@ -80,7 +79,6 @@ class Effektio extends StatelessWidget {
               '/profile': (BuildContext context) => const SocialProfileScreen(),
               '/signup': (BuildContext context) => const SignupScreen(),
               '/gallery': (BuildContext context) => const GalleryScreen(),
-              '/addTodo': (BuildContext context) => const AddToDoScreen(),
             },
           ),
         ),
@@ -155,7 +153,7 @@ class _EffektioHomeState extends State<EffektioHome>
         if (mounted) {
           setState(() {
             if (value.hasAvatar()) {
-              displayAvatar = value.getAvatar();
+              displayAvatar = value.getThumbnail(50, 50);
             }
             displayName = value.getDisplayName();
           });
@@ -302,7 +300,6 @@ class _EffektioHomeState extends State<EffektioHome>
             userId: client.userId().toString(),
             displayName: displayName,
             displayAvatar: displayAvatar,
-            client: client,
           ),
           bottomNavigationBar: TabBar(
             labelColor: AppCommonTheme.primaryColor,
