@@ -1,6 +1,7 @@
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/chat_list_controller.dart';
 import 'package:effektio/controllers/receipt_controller.dart';
+import 'package:effektio/models/JoinedRoom.dart';
 import 'package:effektio/screens/HomeScreens/chat/ChatScreen.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio/widgets/CustomAvatar.dart';
@@ -15,16 +16,12 @@ import 'package:intl/intl.dart';
 
 class ChatListItem extends StatefulWidget {
   final Client client;
-  final Conversation room;
-  final RoomMessage? latestMessage;
-  final List<types.User> typingUsers;
+  final JoinedRoom room;
 
   const ChatListItem({
     Key? key,
     required this.client,
     required this.room,
-    this.latestMessage,
-    required this.typingUsers,
   }) : super(key: key);
 
   @override
@@ -32,35 +29,21 @@ class ChatListItem extends StatefulWidget {
 }
 
 class _ChatListItemState extends State<ChatListItem> {
-  Future<FfiBufferUint8>? avatar;
-  String? displayName;
-  String? userId;
+  final ReceiptController recieptController = Get.find<ReceiptController>();
+  final ChatListController chatListController = Get.find<ChatListController>();
 
   List<Member> activeMembers = [];
 
   @override
   void initState() {
     super.initState();
-
-    widget.room.getProfile().then((value) {
-      if (mounted) {
-        setState(() {
-          if (value.hasAvatar()) {
-            avatar = value.getAvatar();
-          }
-          displayName = value.getDisplayName();
-        });
-      }
-    });
-
-    userId = widget.client.account().userId();
-
+    chatListController.setRoomProfile(widget.room.conversation, widget.room);
     getActiveMembers();
   }
 
   @override
   Widget build(BuildContext context) {
-    String roomId = widget.room.getRoomId();
+    String roomId = widget.room.conversation.getRoomId();
     // ToDo: UnreadCounter
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -69,20 +52,32 @@ class _ChatListItemState extends State<ChatListItem> {
           onTap: () => handleTap(context),
           leading: CustomAvatar(
             uniqueKey: roomId,
-            avatar: avatar,
-            displayName: displayName,
+            avatar: widget.room.avatar,
+            displayName: widget.room.displayName,
             radius: 25,
-            cacheHeight: 120,
-            cacheWidth: 120,
+            cacheHeight: 62,
+            cacheWidth: 60,
             isGroup: true,
             stringName: simplifyRoomId(roomId)!,
           ),
-          title: buildTitle(context),
+          title: _TitleWidget(
+            displayName: widget.room.displayName,
+            context: context,
+          ),
           subtitle: GetBuilder<ChatListController>(
             id: 'chatroom-$roomId-subtitle',
-            builder: (controller) => buildSubtitle(context),
+            builder: (_) => _SubtitleWidget(
+              typingUsers: widget.room.typingUsers,
+              latestMessage: widget.room.latestMessage,
+            ),
           ),
-          trailing: buildTrailing(context),
+          trailing: _TrailingWidget(
+            controller: recieptController,
+            room: widget.room.conversation,
+            latestMessage: widget.room.latestMessage,
+            activeMembers: activeMembers,
+            userId: widget.client.account().userId(),
+          ),
         ),
         const Padding(
           padding: EdgeInsets.only(bottom: 5),
@@ -102,15 +97,30 @@ class _ChatListItemState extends State<ChatListItem> {
       MaterialPageRoute(
         builder: (context) => ChatScreen(
           client: widget.client,
-          room: widget.room,
-          roomName: displayName,
-          roomAvatar: avatar,
+          conversation: widget.room.conversation,
+          name: widget.room.displayName,
+          avatar: widget.room.avatar,
         ),
       ),
     );
   }
 
-  Widget buildTitle(BuildContext context) {
+  Future<void> getActiveMembers() async {
+    activeMembers = (await widget.room.conversation.activeMembers()).toList();
+  }
+}
+
+class _TitleWidget extends StatelessWidget {
+  const _TitleWidget({
+    required this.displayName,
+    required this.context,
+  });
+
+  final String? displayName;
+  final BuildContext context;
+
+  @override
+  Widget build(BuildContext context) {
     if (displayName == null) {
       return Text(
         AppLocalizations.of(context)!.loadingName,
@@ -122,25 +132,35 @@ class _ChatListItemState extends State<ChatListItem> {
       style: ChatTheme01.chatTitleStyle,
     );
   }
+}
 
-  Widget buildSubtitle(BuildContext context) {
-    if (widget.typingUsers.isNotEmpty) {
+class _SubtitleWidget extends StatelessWidget {
+  const _SubtitleWidget({
+    required this.typingUsers,
+    required this.latestMessage,
+  });
+  final List<types.User> typingUsers;
+  final RoomMessage? latestMessage;
+
+  @override
+  Widget build(BuildContext context) {
+    if (typingUsers.isNotEmpty) {
       return Container(
         margin: const EdgeInsets.symmetric(vertical: 10),
         child: Text(
-          getUserPlural(widget.typingUsers),
+          getUserPlural(typingUsers),
           style: ChatTheme01.latestChatStyle.copyWith(
             fontStyle: FontStyle.italic,
           ),
         ),
       );
     }
-    if (widget.latestMessage == null) {
-      return const SizedBox();
+    if (latestMessage == null) {
+      return const SizedBox.shrink();
     }
-    RoomEventItem? eventItem = widget.latestMessage!.eventItem();
+    RoomEventItem? eventItem = latestMessage!.eventItem();
     if (eventItem == null) {
-      return const SizedBox();
+      return const SizedBox.shrink();
     }
 
     String sender = eventItem.sender();
@@ -215,29 +235,55 @@ class _ChatListItemState extends State<ChatListItem> {
           ],
         );
     }
-
-    // exclude state event
-    return const SizedBox();
+    return const SizedBox.shrink();
   }
 
-  Widget? buildTrailing(BuildContext context) {
-    if (widget.latestMessage == null) {
-      return null;
+  String getUserPlural(List<types.User> authors) {
+    if (authors.isEmpty) {
+      return '';
+    } else if (authors.length == 1) {
+      return '${authors[0].firstName} is typing...';
+    } else if (authors.length == 2) {
+      return '${authors[0].firstName} and ${authors[1].firstName} is typing...';
+    } else {
+      return '${authors[0].firstName} and ${authors.length - 1} others typing...';
     }
-    RoomEventItem? eventItem = widget.latestMessage!.eventItem();
+  }
+}
+
+class _TrailingWidget extends StatelessWidget {
+  const _TrailingWidget({
+    required this.controller,
+    required this.room,
+    required this.activeMembers,
+    this.latestMessage,
+    required this.userId,
+  });
+  final ReceiptController controller;
+  final Conversation room;
+  final List<Member> activeMembers;
+  final RoomMessage? latestMessage;
+  final String? userId;
+
+  @override
+  Widget build(BuildContext context) {
+    if (latestMessage == null) {
+      return const SizedBox.shrink();
+    }
+    RoomEventItem? eventItem = latestMessage!.eventItem();
     if (eventItem == null) {
-      return null;
+      return const SizedBox.shrink();
     }
     String senderID = '';
     types.Status? messageStatus;
     int ts = eventItem.originServerTs();
-    var receiptController = Get.find<ReceiptController>();
-    List<String> seenByList = receiptController.getSeenByList(
-      widget.room.getRoomId(),
+
+    List<String> seenByList = controller.getSeenByList(
+      room.getRoomId(),
       ts,
     );
 
-    senderID = widget.latestMessage!.eventItem()!.sender();
+    senderID = latestMessage!.eventItem()!.sender();
 
     messageStatus = seenByList.isEmpty
         ? types.Status.sent
@@ -254,13 +300,22 @@ class _ChatListItemState extends State<ChatListItem> {
           style: ChatTheme01.latestChatDateStyle,
         ),
         senderID == userId
-            ? customStatusBuilder(messageStatus)
-            : const SizedBox(),
+            ? _CustomStatusWidget(status: messageStatus)
+            : const SizedBox.shrink(),
       ],
     );
   }
+}
 
-  Widget customStatusBuilder(types.Status status) {
+class _CustomStatusWidget extends StatelessWidget {
+  const _CustomStatusWidget({
+    required this.status,
+  });
+
+  final types.Status status;
+
+  @override
+  Widget build(BuildContext context) {
     if (status == types.Status.delivered) {
       return SvgPicture.asset('assets/images/deliveredIcon.svg');
     } else if (status == types.Status.seen) {
@@ -283,21 +338,5 @@ class _ChatListItemState extends State<ChatListItem> {
         height: 12,
       );
     }
-  }
-
-  String getUserPlural(List<types.User> authors) {
-    if (authors.isEmpty) {
-      return '';
-    } else if (authors.length == 1) {
-      return '${authors[0].firstName} is typing...';
-    } else if (authors.length == 2) {
-      return '${authors[0].firstName} and ${authors[1].firstName} is typing...';
-    } else {
-      return '${authors[0].firstName} and ${authors.length - 1} others typing...';
-    }
-  }
-
-  Future<void> getActiveMembers() async {
-    activeMembers = (await widget.room.activeMembers()).toList();
   }
 }
