@@ -3,12 +3,12 @@ import 'package:effektio/controllers/todo_controller.dart';
 import 'package:effektio/models/Team.dart';
 import 'package:effektio/models/ToDoList.dart';
 import 'package:effektio/models/ToDoTask.dart';
-import 'package:effektio/screens/HomeScreens/todo/screens/CommentsScreen.dart';
 import 'package:effektio/widgets/AddTaskDialog.dart';
 import 'package:effektio/widgets/ExpandableText.dart';
 import 'package:effektio/widgets/ToDoTaskView.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
 
 class ToDoListView extends StatelessWidget {
@@ -54,10 +54,10 @@ class TodoCard extends StatelessWidget {
     required this.isExpanded,
     required this.expandBtn,
   });
+  final ToDoController controller;
   final int index;
   final bool isExpanded;
   final bool expandBtn;
-  final ToDoController controller;
   @override
   Widget build(BuildContext context) {
     return Card(
@@ -75,153 +75,57 @@ class TodoCard extends StatelessWidget {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
-            buildHeaderContent(
-              controller.todos[index].name,
-              controller.todos[index].team,
+            _HeaderWidget(
+              title: controller.todos[index].name,
+              team: controller.todos[index].team,
             ),
-            buildDescription(controller.todos[index].description),
-            buildDivider(),
-            buildComments(context),
-            buildDivider(),
-            buildTasksRatio(controller.todos[index]),
-            buildTasksSection(
-              context,
-              controller.todos[index],
+            _DescriptionWidget(
+              description: controller.todos[index].description,
+            ),
+            const _DividerWidget(),
+            _CommentsWidget(context: context),
+            const _DividerWidget(),
+            _TasksRatioWidget(
+              controller: controller,
+              index: index,
               isExpanded: isExpanded,
-              expandBtn: expandBtn,
-            )
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget buildHeaderContent(String title, Team? team) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 5),
-      child: Row(
-        children: [
-          Flexible(
-            child: Text(
-              title,
-              style: ToDoTheme.listTitleTextStyle,
+              todo: controller.todos[index],
             ),
-          ),
-          const SizedBox(
-            width: 8.0,
-          ),
-          team != null
-              ? Container(
-                  padding: const EdgeInsets.all(8.0),
-                  decoration: const BoxDecoration(
-                    color: AppCommonTheme.secondaryColor,
-                    borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                  ),
-                  child: Text(
-                    team.name!,
-                    style: ToDoTheme.listTagTextStyle,
-                  ),
-                )
-              : const SizedBox(),
-        ],
-      ),
-    );
-  }
-
-  Widget buildDescription(String? description) {
-    return SizedBox(
-      child: (description != null || description!.isNotEmpty)
-          ? description.length > 80
-              ? ExpandableText(description)
-              : Text(description, style: ToDoTheme.descriptionTextStyle)
-          : const SizedBox.shrink(),
-    );
-  }
-
-  Widget buildDivider() {
-    return const Divider(
-      color: ToDoTheme.listDividerColor,
-      indent: 0,
-      endIndent: 0,
-      thickness: 1,
-    );
-  }
-
-  Widget buildTasksRatio(ToDoList todo) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: <Widget>[
-        const Text(
-          'Task: ',
-          style: ToDoTheme.listSubtitleTextStyle,
-        ),
-        Text(
-          '${controller.getCompletedTasks(todo)}/${todo.tasks.length} completed',
-          style: ToDoTheme.listSubtitleTextStyle.copyWith(
-            color: ToDoTheme.calendarColor,
-          ),
-        ),
-        const Spacer(),
-        IconButton(
-          onPressed: () => controller.toggleCardExpand(index, isExpanded),
-          icon: Icon(
-            isExpanded
-                ? FlutterIcons.ios_arrow_up_ion
-                : FlutterIcons.ios_arrow_down_ion,
-          ),
-          color: ToDoTheme.primaryTextColor,
-        ),
-      ],
-    );
-  }
-
-  Widget buildComments(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Row(
-          children: [
-            const Icon(
-              FlutterIcons.heart_evi,
-              color: ToDoTheme.primaryTextColor,
-            ),
-            const SizedBox(
-              width: 8,
-            ),
-            InkWell(
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => const ToDoCommentScreen(),
-                  ),
-                );
-              },
-              child: Row(
-                children: const [
-                  Icon(
-                    FlutterIcons.comment_evi,
-                    color: ToDoTheme.primaryTextColor,
-                  ),
-                ],
+            Obx(
+              () => _TasksWidget(
+                controller: controller,
+                index: index,
+                context: context,
+                todo: controller.todos[index],
+                isExpanded: isExpanded,
+                expandBtn: expandBtn,
               ),
             )
           ],
         ),
-        const Icon(
-          FlutterIcons.bookmark_border_mdi,
-          color: ToDoTheme.primaryTextColor,
-        ),
-      ],
+      ),
     );
   }
+}
 
-  Widget buildTasksSection(
-    BuildContext context,
-    ToDoList todo, {
-    required isExpanded,
-    required expandBtn,
-  }) {
+class _TasksWidget extends StatelessWidget {
+  const _TasksWidget({
+    required this.controller,
+    required this.index,
+    required this.context,
+    required this.todo,
+    required this.isExpanded,
+    required this.expandBtn,
+  });
+  final ToDoController controller;
+  final int index;
+  final BuildContext context;
+  final ToDoList todo;
+  final bool isExpanded;
+  final bool expandBtn;
+
+  @override
+  Widget build(BuildContext context) {
     List<ToDoTask> _pendingTasks =
         todo.tasks.where((element) => element.progressPercent < 100).toList();
     List<ToDoTask> _completedTasks =
@@ -237,7 +141,6 @@ class TodoCard extends StatelessWidget {
               _pendingTasks.length,
               (index) => ToDoTaskView(
                 task: _pendingTasks[index],
-                controller: controller,
                 todoList: todo,
               ),
             ),
@@ -295,7 +198,6 @@ class TodoCard extends StatelessWidget {
                           type: MaterialType.transparency,
                           child: AddTaskDialogBox(
                             toDoList: todo,
-                            controller: controller,
                           ),
                         ),
                       );
@@ -317,13 +219,166 @@ class TodoCard extends StatelessWidget {
                   _completedTasks.length,
                   (index) => ToDoTaskView(
                     task: _completedTasks[index],
-                    controller: controller,
                     todoList: todo,
                   ),
                 ),
               ),
             ),
           ),
+        ],
+      ),
+    );
+  }
+}
+
+class _TasksRatioWidget extends StatelessWidget {
+  const _TasksRatioWidget({
+    required this.controller,
+    required this.index,
+    required this.isExpanded,
+    required this.todo,
+  });
+  final ToDoController controller;
+  final int index;
+  final bool isExpanded;
+  final ToDoList todo;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: <Widget>[
+        const Text(
+          'Task: ',
+          style: ToDoTheme.listSubtitleTextStyle,
+        ),
+        Text(
+          '${controller.getCompletedTasks(todo)}/${todo.tasks.length} completed',
+          style: ToDoTheme.listSubtitleTextStyle.copyWith(
+            color: ToDoTheme.calendarColor,
+          ),
+        ),
+        const Spacer(),
+        IconButton(
+          onPressed: () => controller.toggleCardExpand(index, isExpanded),
+          icon: Icon(
+            isExpanded
+                ? FlutterIcons.ios_arrow_up_ion
+                : FlutterIcons.ios_arrow_down_ion,
+          ),
+          color: ToDoTheme.primaryTextColor,
+        ),
+      ],
+    );
+  }
+}
+
+class _CommentsWidget extends StatelessWidget {
+  const _CommentsWidget({
+    required this.context,
+  });
+
+  final BuildContext context;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.start,
+      children: [
+        const Icon(
+          FlutterIcons.heart_evi,
+          color: ToDoTheme.primaryTextColor,
+        ),
+        Padding(
+          padding: const EdgeInsets.only(left: 8, top: 3),
+          child: InkWell(
+            onTap: () {},
+            child: SvgPicture.asset(
+              'assets/images/message.svg',
+              color: Colors.white,
+              height: 18,
+              width: 18,
+            ),
+          ),
+        ),
+        const Spacer(),
+        const Icon(
+          FlutterIcons.bookmark_border_mdi,
+          color: ToDoTheme.primaryTextColor,
+        ),
+      ],
+    );
+  }
+}
+
+class _DescriptionWidget extends StatelessWidget {
+  const _DescriptionWidget({
+    required this.description,
+  });
+
+  final String? description;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      child: (description != null || description!.isNotEmpty)
+          ? description!.length > 80
+              ? ExpandableText(description!)
+              : Text(description!, style: ToDoTheme.descriptionTextStyle)
+          : const SizedBox.shrink(),
+    );
+  }
+}
+
+class _DividerWidget extends StatelessWidget {
+  const _DividerWidget();
+
+  @override
+  Widget build(BuildContext context) {
+    return const Divider(
+      color: ToDoTheme.listDividerColor,
+      indent: 0,
+      endIndent: 0,
+      thickness: 1,
+    );
+  }
+}
+
+class _HeaderWidget extends StatelessWidget {
+  const _HeaderWidget({
+    required this.title,
+    required this.team,
+  });
+
+  final String title;
+  final Team? team;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 5),
+      child: Row(
+        children: [
+          Text(
+            title,
+            style: ToDoTheme.listTitleTextStyle,
+          ),
+          const SizedBox(
+            width: 8.0,
+          ),
+          team != null
+              ? Container(
+                  padding: const EdgeInsets.all(8.0),
+                  decoration: const BoxDecoration(
+                    color: AppCommonTheme.secondaryColor,
+                    borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                  ),
+                  child: Text(
+                    team!.name!,
+                    style: ToDoTheme.listTagTextStyle,
+                  ),
+                )
+              : const SizedBox.shrink(),
         ],
       ),
     );
