@@ -2,49 +2,70 @@ import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/todo_controller.dart';
 import 'package:effektio/models/ToDoList.dart';
 import 'package:effektio/models/ToDoTask.dart';
+import 'package:effektio/screens/HomeScreens/todo/TaskDetail.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:get/get.dart';
 import 'package:intl/intl.dart';
 
-class ToDoTaskView extends StatelessWidget {
+class ToDoTaskView extends StatefulWidget {
   final ToDoTask task;
   final ToDoList todoList;
-  final ToDoController controller;
   const ToDoTaskView({
     Key? key,
     required this.task,
     required this.todoList,
-    required this.controller,
   }) : super(key: key);
+
+  @override
+  State<ToDoTaskView> createState() => _ToDoTaskViewState();
+}
+
+class _ToDoTaskViewState extends State<ToDoTaskView> {
+  final ToDoController controller = Get.find<ToDoController>();
+  late int idx;
+  late int listIdx;
+  @override
+  void initState() {
+    super.initState();
+    listIdx = controller.todos.indexOf(widget.todoList);
+    idx = widget.todoList.tasks.indexOf(widget.task);
+  }
+
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: () => {},
+      onTap: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+            builder: (context) => TaskDetailScreen(
+              index: idx,
+              listIndex: listIdx,
+            ),
+          ),
+        );
+      },
       child: TaskCard(
-        task: task,
         controller: controller,
-        todoList: todoList,
+        task: widget.task,
+        todoList: widget.todoList,
       ),
     );
   }
 }
 
-class TaskCard extends StatefulWidget {
+class TaskCard extends StatelessWidget {
   const TaskCard({
     super.key,
+    required this.controller,
     required this.task,
     required this.todoList,
-    required this.controller,
   });
+  final ToDoController controller;
   final ToDoTask task;
   final ToDoList todoList;
-  final ToDoController controller;
-  @override
-  State<TaskCard> createState() => _TaskCardState();
-}
-
-class _TaskCardState extends State<TaskCard> {
   @override
   Widget build(BuildContext context) {
     return Card(
@@ -61,8 +82,8 @@ class _TaskCardState extends State<TaskCard> {
               children: <Widget>[
                 Flexible(
                   child: InkWell(
-                    onTap: () async => await widget.controller
-                        .markToDoTask(widget.task, widget.todoList)
+                    onTap: () async => await controller
+                        .updateToDoTask(task, todoList, null, null, null)
                         .then((res) => debugPrint('TOGGLE CHECK')),
                     child: CircleAvatar(
                       backgroundColor: AppCommonTheme.transparentColor,
@@ -71,7 +92,7 @@ class _TaskCardState extends State<TaskCard> {
                         height: 25,
                         width: 25,
                         decoration: BoxDecoration(
-                          color: (widget.task.progressPercent >= 100)
+                          color: (task.progressPercent >= 100)
                               ? ToDoTheme.activeCheckColor
                               : ToDoTheme.inactiveCheckColor,
                           shape: BoxShape.circle,
@@ -80,7 +101,7 @@ class _TaskCardState extends State<TaskCard> {
                             color: ToDoTheme.floatingABColor,
                           ),
                         ),
-                        child: checkBuilder(),
+                        child: _CheckWidget(task: task),
                       ),
                     ),
                   ),
@@ -89,9 +110,9 @@ class _TaskCardState extends State<TaskCard> {
                   child: Padding(
                     padding: const EdgeInsets.only(left: 8),
                     child: Text(
-                      widget.task.name,
+                      task.name,
                       style: ToDoTheme.taskTitleTextStyle.copyWith(
-                        decoration: (widget.task.progressPercent >= 100)
+                        decoration: (task.progressPercent >= 100)
                             ? TextDecoration.lineThrough
                             : null,
                       ),
@@ -105,7 +126,7 @@ class _TaskCardState extends State<TaskCard> {
           Padding(
             padding: const EdgeInsets.all(10),
             child: Row(
-              mainAxisAlignment: widget.task.progressPercent >= 100
+              mainAxisAlignment: task.progressPercent >= 100
                   ? MainAxisAlignment.end
                   : MainAxisAlignment.start,
               children: [
@@ -113,7 +134,7 @@ class _TaskCardState extends State<TaskCard> {
                   padding:
                       const EdgeInsets.symmetric(horizontal: 7, vertical: 5),
                   child: Visibility(
-                    visible: widget.task.progressPercent >= 100,
+                    visible: task.progressPercent >= 100,
                     child: const Icon(
                       FlutterIcons.ios_checkmark_circle_outline_ion,
                       color: Colors.red,
@@ -121,26 +142,58 @@ class _TaskCardState extends State<TaskCard> {
                     ),
                   ),
                 ),
-                SvgPicture.asset(
-                  'assets/images/calendar-2.svg',
-                  color: widget.task.progressPercent >= 100
-                      ? Colors.red
-                      : ToDoTheme.todayCalendarColor,
+                task.due != null
+                    ? SvgPicture.asset(
+                        'assets/images/calendar-2.svg',
+                        color: task.progressPercent >= 100
+                            ? Colors.red
+                            : ToDoTheme.todayCalendarColor,
+                      )
+                    : const SizedBox.shrink(),
+                Padding(
+                  padding: const EdgeInsets.only(left: 8),
+                  child: task.due != null
+                      ? Text(
+                          task.progressPercent >= 100
+                              ? DateFormat('H:mm E, d MMM')
+                                  .format(task.due!.toUtc())
+                              : DateFormat('E, d MMM')
+                                  .format(task.due!.toUtc()),
+                          style: task.progressPercent >= 100
+                              ? ToDoTheme.todayCalendarTextStyle
+                                  .copyWith(color: Colors.red)
+                              : ToDoTheme.todayCalendarTextStyle,
+                        )
+                      : const SizedBox.shrink(),
                 ),
                 Padding(
                   padding: const EdgeInsets.only(left: 8),
-                  child: Text(
-                    widget.task.progressPercent >= 100
-                        ? DateFormat('H:m E, d MMM')
-                            .format(widget.task.due!.toUtc())
-                        : DateFormat('E, d MMM')
-                            .format(widget.task.due!.toUtc()),
-                    style: widget.task.progressPercent >= 100
-                        ? ToDoTheme.todayCalendarTextStyle
-                            .copyWith(color: Colors.red)
-                        : ToDoTheme.todayCalendarTextStyle,
+                  child: Visibility(
+                    visible: task.commentsManager.hasComments(),
+                    child: Row(
+                      children: <Widget>[
+                        const Icon(
+                          FlutterIcons.dot_single_ent,
+                          color: Colors.grey,
+                        ),
+                        SvgPicture.asset(
+                          'assets/images/message.svg',
+                          color: Colors.white,
+                          height: 18,
+                          width: 18,
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(left: 8),
+                          child: Text(
+                            '${task.commentsManager.commentsCount()}',
+                            style: ToDoTheme.todayCalendarTextStyle
+                                .copyWith(color: Colors.white),
+                          ),
+                        )
+                      ],
+                    ),
                   ),
-                ),
+                )
               ],
             ),
           ),
@@ -148,10 +201,19 @@ class _TaskCardState extends State<TaskCard> {
       ),
     );
   }
+}
 
-  Widget? checkBuilder() {
-    if ((widget.task.progressPercent < 100)) {
-      return null;
+class _CheckWidget extends StatelessWidget {
+  const _CheckWidget({
+    required this.task,
+  });
+
+  final ToDoTask task;
+
+  @override
+  Widget build(BuildContext context) {
+    if ((task.progressPercent < 100)) {
+      return const SizedBox.shrink();
     }
     return const Icon(
       Icons.done_outlined,
